Don't provide AXDescriptions that do not make sense
The window buttons are described sufficiently by their subrole (i.e. "zoom button").

The window is described sufficiently and properly by its AXTitle (which is name
of the movie). Providing a non-empty AXDescription does not make sense and
replaces the name of the movie in window title (notable e.g. with VO-F2).